fucntionality for deleting users; refactoring

+++ app/imports/api/standards-types/methods.js
@@ -4,7 +4,7 @@ import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import StandardsTypeService from './standards-type-service.js';
 import { StandardsTypeSchema } from './standards-type-schema.js';
 import { IdSchema } from '../schemas.js';
-import { checkUserId } from '../checkers.js';
+
 
 export const insert = new ValidatedMethod({
   name: 'StandardsTypes.insert',
@@ -12,9 +12,11 @@ export const insert = new ValidatedMethod({
   validate: StandardsTypeSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot create a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot create a standard type'
+      );
+    }
 
     return StandardsTypeService.insert(doc);
   }
@@ -29,9 +31,11 @@ export const update = new ValidatedMethod({
   }]).validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot update a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot update a standard type'
+      );
+    }
 
     return StandardsTypeService.update(doc);
   }
@@ -43,9 +47,11 @@ export const remove = new ValidatedMethod({
   validate: IdSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot remove a standard type'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot remove a standard type'
+      );
+    }
 
     return StandardsTypeService.remove(doc);
   }
